deps/sh3dkit/requirements.txt
skia-python==87.8
svg.py==1.6.0
sh3d.py>=0.2.1
click==8.1.8